Add TextualConflictFilter and fix tests
