+++ src/test/integration/FileTestModule.groovy
@@ -6,18 +6,14 @@ import interfaces.CommitFilter
 import interfaces.DataCollector
 import interfaces.OutputProcessor
 import interfaces.ProjectProcessor
-import services.commitFilters.InCommitListAndHasMutuallyModifiedMethodsFilter
+import services.commitFilters.InCommitListMutuallyModifiedMethodsTextualConflictCommitFilter
 import services.dataCollectors.BuildRequester
 import services.dataCollectors.MergeConflictCollector
 import services.dataCollectors.StatisticsCollector
 import services.dataCollectors.modifiedLinesCollector.ModifiedLinesCollector
-import services.outputProcessors.FetchBuildsOutputProcessor
 import services.outputProcessors.GenerateSootInputFilesOutputProcessor
 import services.outputProcessors.soot.RunSootAnalysisOutputProcessor
 import services.projectProcessors.FilterNonExistentProjectsProcessor
-import services.projectProcessors.ForkAndEnableTravisProcessor
-import util.EmptyOutputProcessor
-import util.EmptyProjectProcessor
 
 public class FileTestModule extends AbstractModule {
 
@@ -25,12 +21,11 @@ public class FileTestModule extends AbstractModule {
     protected void configure() {
         Multibinder<DataCollector> dataCollectorBinder = Multibinder.newSetBinder(binder(), DataCollector.class)
 
-        dataCollectorBinder.addBinding().to(MergeConflictCollector.class)
         dataCollectorBinder.addBinding().to(ModifiedLinesCollector.class)
         dataCollectorBinder.addBinding().to(StatisticsCollector.class)
         dataCollectorBinder.addBinding().to(BuildRequester.class)
 
-        bind(CommitFilter.class).to(InCommitListAndHasMutuallyModifiedMethodsFilter.class)
+        bind(CommitFilter.class).to(InCommitListMutuallyModifiedMethodsTextualConflictCommitFilter.class)
 
         Multibinder<ProjectProcessor> projectProcessorBinder = Multibinder.newSetBinder(binder(), ProjectProcessor.class)
         projectProcessorBinder.addBinding().to(FilterNonExistentProjectsProcessor.class)
