+++ src/test/integration/fileTest/compare/statistics/result.csv
@@ -1,8 +1,5 @@
-jsoup,ea3225714c9e6e02ab17c4a79a88b71d1ea6eae3,false,68.0,1042.0,2.0,30.59411708155671,1.0,0
 jsoup,7ba0ee75547878ddccc5ab50777b67c712aa73e4,false,65.99242380758567,999.4718605343525,12.84523257866513,247.39846402110098,9.16515138991168,19
-jsoup,e38b916ba1ebd3f5c00a6b6b8e77a4cdfea2800f,false,58.48076606885378,922.489024324951,19.078784028338912,240.70313666423212,14.0,0
 jsoup,a8b6982de98ff76ef254031d7152fff57f6bf941,false,52.49761899362675,865.4825243758536,3.0,31.74901573277509,31.796226191169293,341
-jsoup,fee4762322f85a1109edd75ccb67f38cf5008c80,false,38.0,715.9748598938373,10.392304845413264,251.76179217665256,3.7416573867739413,13
 jsoup,3f7d2c71dbbbb289c684f339874eed8ac2747fa0,false,27.386127875258307,642.8880151317179,4.69041575982343,68.05879810869422,359.76520120767657,26
 jsoup,821b021a1e5fc75719c71ae5ea84157d770c1a9d,false,25.0,635.4761993969562,6.48074069840786,53.665631459994955,370.994609125254,4
 jsoup,a44e18aa3c1fcd25a68a5965f9490d8f7d026509,false,15.0,549.9772722576815,8.0,101.83319694480774,33.76388603226827,377
